[refactor] #32 - 알림창 ui 수정 및 패키지 구조 수정
